refactor: update Colors import paths and improve message formatting in parkour commands
